requirements.txt
# Core Framework
fastapi==0.116.2
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv

# Google Cloud & AI
google-cloud-aiplatform
google-cloud-firestore
google-cloud-storage
google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2

# Firebase
firebase-admin

# Research Paper Sources
arxiv==2.2.0
biopython
scholarly

# Web Scraping & APIs
requests
beautifulsoup4
lxml
httpx
aiohttp
google-search-results

# PDF Generation
reportlab
PyPDF2
fpdf2==2.8.4
pillow
matplotlib

# Data Processing
numpy
pandas
scipy

# Database & Caching
redis
motor
pymongo

# Utilities
feedparser==6.0.12
python-multipart
email-validator
